requirements.txt
pandas>=2.2
numpy>=1.26
pyarrow>=14.0
scikit-learn>=1.4
lightgbm>=4.3
xgboost>=2.0
catboost>=1.2
category-encoders>=2.6
joblib>=1.3
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
ipykernel>=6.29
pytest>=7.0
